Added VueJs,CartService,OrderService

+++ database/migrations/2024_08_12_160530_add_user_id_to_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id');
+            $table->integer('user_id')->after('id');
         });
     }
